// types.ts
export interface SubStatusConfig {
    name: string;
    note: string;
}

export interface ProjectStatusConfig {
    id: string;
    name: string;
    color: string; // hex color
    subStatuses: SubStatusConfig[];
    note: string;
}
// ...
export interface PhysicalItem {
    name: string;
    price: number;
}

export interface Package {
  id: string;
  name: string;
  price: number;
  physicalItems: PhysicalItem[];
  digitalItems: string[];
  processingTime: string;
  defaultPrintingCost?: number;
  defaultTransportCost?: number;
  photographers?: string;
  videographers?: string;
}
// ...
export interface NotificationSettings {
    newProject: boolean;
    paymentConfirmation: boolean;
    deadlineReminder: boolean;
}

export interface SecuritySettings {
    twoFactorEnabled: boolean;
}

export interface Profile {
    fullName: string;
    email: string;
    phone: string;
    companyName: string;
    website: string;
    address: string;
    bankAccount: string;
    authorizedSigner: string;
    idNumber?: string;
    bio: string;
    incomeCategories: string[];
    expenseCategories: string[];
    projectTypes: string[];
    eventTypes: string[];
    assetCategories: string[];
    sopCategories: string[];
    projectStatusConfig: ProjectStatusConfig[];
    notificationSettings: NotificationSettings;
    securitySettings: SecuritySettings;
    briefingTemplate: string;
    termsAndConditions?: string;
    contractTemplate?: string;
}